recipes-initscripts/cml-boot/cml-boot: TPM initialisation for RPi5
RPi does not boot using UEFI, which is why we fill the PCRs with values
from the RPi OTP for now, namely the serial number and the device secret
key.

TODO: This should actually be implemented in meta-gyroidos-rpi, which is
complicated because we still need the spacer #RPI_TPM_INIT# here. This
will be fixed by an upcoming modular rewrite of the init script
build process.

TODO: Measure device UUID to PCR1 but i have not found out yet how it
can be obtained using vcmailbox.

Signed-off-by: Johannes Wiesboeck <johannes.wiesboeck@aisec.fraunhofer.de>